Force all bits on GPIO0 pins to 0 for testing.
This way we can connect an osiliscope or logic analyzer to any of these pins
and trigger on any pulse wider than 550ns to see glitches. A 0 bit is
normally ~300ns and a 1 bit is anything over 550ns, so if we see one of
these glitches when sending all 0 bits then we would have had an LED flash
since the 0 bit was glitched to a 1 bit.

Well I did this very test and it only took a few seconds to get a glitch, and
the glitch also happened on GPIO1 pins that happened to be sending 0 bits also.

This test 100% confirms that the problem is happenig at the PRU to GPIO level
so now we can move forward, and we will know if we fixed it if we runt this
same test again and are not able to get a glitch.